requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==2.1.1
Django==4.2.15
django-environ==0.11.2
django-single-session==0.2.0
django-utils-six==2.0
djangorestframework==3.15.2
idna==3.7
psycopg2-binary==2.9.3
pytz==2024.1
requests==2.28.1
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==1.26.19
